WebMQTT/flask_app/requirements.txt
hbmqtt
Flask==2.2.2
flask-cors==3.0.10
paho-mqtt==1.6.1
websockets==8.1
Werkzeug==2.2.2
opencv-python
flask-socketio